ApiRequester: Extract processing code to an abstract base.
The only thing the baseclass requires in a derived one is the actual message
exchange. The existing ApiRequester uses Guzzle to make a call via network to
an actual HTTP server, but keeping the message in-proc is also an option.
